add gender to user, add a validation for user creation

+++ service/src/main/java/com/reimbes/implementation/AdminServiceImpl.java
@@ -2,6 +2,7 @@ package com.reimbes.implementation;
 
 import com.reimbes.AdminService;
 import com.reimbes.ReimsUser;
+import com.reimbes.exception.DataConstraintException;
 import com.reimbes.exception.NotFoundException;
 import com.reimbes.exception.ReimsException;
 import org.slf4j.Logger;
@@ -10,6 +11,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -51,12 +53,17 @@ public class AdminServiceImpl implements AdminService {
 
     @Override
     public ReimsUser createUser(ReimsUser user) throws ReimsException {
+        if (user.getRole() == ReimsUser.Role.USER && user.getGender() == null)
+            throw new DataConstraintException("NULL_ATTRIBUTE_GENDER");
         return userService.create(user);
     }
 
     @Override
-    public Object updateUser(long id, ReimsUser user, HttpServletResponse response) throws ReimsException {
+    public ReimsUser updateUser(long id, ReimsUser user, HttpServletResponse response) throws ReimsException {
         ReimsUser currentUser = userService.getUserByUsername(utils.getUsername());
+
+        if (user.getRole() == ReimsUser.Role.USER && user.getGender() == null)
+            throw new DataConstraintException("NULL_ATTRIBUTE_GENDER");
 
         // if admin try to update his data
         if (currentUser.getId() == id) return userService.updateMyData(user, response);
@@ -65,7 +72,9 @@ public class AdminServiceImpl implements AdminService {
     }
 
     @Override
-    public void deleteUser(long id) {
+    public void deleteUser(long id) throws ReimsException{
+        ReimsUser currentUser = userService.getUserByUsername(utils.getUsername());
+        if (currentUser.getId() == id) throw new ReimsException("SELF_DELETION", HttpStatus.METHOD_NOT_ALLOWED, 405);
         userService.deleteUser(id);
     }
 }
